Fix state parse failure on fractional XP (SkillState.xp must be double)
The client mirrored SkillState.xp as int, but server XP is fractional (mining
17.5/ore -> 157.5 total; combat hitpoints XP = dmg×1.33). glaze then failed the
whole state message with parse_number_failure on the non-integer value, dropping
every tick while mining (and intermittently in combat). Changed xp to double and
cast to int at the display sites. Also kept a detailed glz::format_error in the
state-parse failure log for future debugging.

+++ client-native/src/shared/SharedTypes.hpp
@@ -94,7 +94,10 @@ struct ItemStack {
 // SkillsState in TS is Record<SkillId, SkillState>. We deserialize as a
 // flat map so any future skill the server adds shows up automatically.
 struct SkillState {
-  int xp    = 0;
+  // Server XP is fractional (e.g. mining 17.5/ore, combat hitpoints = dmg×1.33),
+  // so this must be floating point — an int here fails JSON parsing on any
+  // non-integer total ("parse_number_failure").
+  double xp = 0;
   int level = 1;
 };
